Add activate\deactivate, edit, dummy play for quiz.

+++ src/Controller/UsersController.php
@@ -26,26 +26,20 @@ class UsersController extends AbstractController
         $repository = $entityManager->getRepository(User::class);
         $users = $repository->findAll();
 
-        $form = $this->createForm(SelectUsersType::class, null, [
-            'users' => $users,
         $paginationUsers = $paginator->paginate(
             $users,
             $request->query->getInt('page', 1),
             10
         );
-        $form = $this->createForm(SelectFormType::class, null, [
+        $form = $this->createForm(SelectUsersType::class, null, [
             'users' => $paginationUsers,
         ]);
+
         return $this->render('users/users_panel.html.twig', [
             'users' => $paginationUsers,
             'selectForm' => $form->createView()
         ]);
     }
-//        if ($form->isSubmitted() && $form->isValid()) {
-//            return $this->forward('App\Controller\UsersController', [
-//                'selectedUsers' => $form->get('selectedUsers')
-//            ]);
-//        }
 
     /**
      * @Route("/admin/users/block", name="app_block")
@@ -55,13 +49,18 @@ class UsersController extends AbstractController
     public function block(Request $request) : Response
     {
         $entityManager = $this->getDoctrine()->getManager();
-        $temp = $request->get('select_form');
-        $usersToBeBlocked = $temp["selectedUsers"];
-        foreach ($usersToBeBlocked as $id) {
-            $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
-            $user->setIsBlocked(true);
+        $form = $this->createForm(SelectUsersType::class, null);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $usersToBeBlocked = $form->getData()['selectedUsers'];
+            foreach ($usersToBeBlocked as $id) {
+                $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+                $user->setIsBlocked(true);
+            }
             $entityManager->flush();
         }
+
         return $this->redirectToRoute('app_users');
     }
 
@@ -73,12 +72,15 @@ class UsersController extends AbstractController
     public function unblock(Request $request) : Response
     {
         $entityManager = $this->getDoctrine()->getManager();
-        $temp = $request->get('select_form');
-        $usersToBeUnblocked = $temp["selectedUsers"];
+        $form = $this->createForm(SelectUsersType::class, null);
+        $form->handleRequest($request);
 
-        foreach ($usersToBeUnblocked as $id) {
-            $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
-            $user->setIsBlocked(false);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $usersToBeUnblocked = $form->getData()['selectedUsers'];
+            foreach ($usersToBeUnblocked as $id) {
+                $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+                $user->setIsBlocked(false);
+            }
             $entityManager->flush();
         }
 
@@ -93,13 +95,18 @@ class UsersController extends AbstractController
     public function delete(Request $request) : Response
     {
         $entityManager = $this->getDoctrine()->getManager();
-        $temp = $request->get('select_form');
-        $usersToBeDeleted = $temp["selectedUsers"];
-        foreach ($usersToBeDeleted as $id) {
-            $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
-            $entityManager->remove($user);
+        $form = $this->createForm(SelectUsersType::class, null);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $usersToBeDeleted = $form->getData()['selectedUsers'];
+            foreach ($usersToBeDeleted as $id) {
+                $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+                $entityManager->remove($user);
+            }
             $entityManager->flush();
         }
+
         return $this->redirectToRoute('app_users');
     }
 
@@ -111,13 +118,18 @@ class UsersController extends AbstractController
     public function activate(Request $request) : Response
     {
         $entityManager = $this->getDoctrine()->getManager();
-        $temp = $request->get('select_form');
-        $usersToBeDeleted = $temp["selectedUsers"];
-        foreach ($usersToBeDeleted as $id) {
-            $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
-            $user->setIsVerified(true);
+        $form = $this->createForm(SelectUsersType::class, null);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $usersToBeActivated = $form->getData()['selectedUsers'];
+            foreach ($usersToBeActivated as $id) {
+                $user = $entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+                $user->setIsVerified(true);
+            }
             $entityManager->flush();
         }
+
         return $this->redirectToRoute('app_users');
     }
 }
